cp action type file
